Fix a crash on 32bit systems
The login-frequency property of the User interface is defined
as uint64 type. Therefore, when setting it with g_object_set,
a 64bit value is collected, and passing an int might lead to
bad things.
https://bugs.freedesktop.org/show_bug.cgi?id=57343

+++ src/daemon.c
@@ -207,7 +207,7 @@ daemon_local_user_is_excluded (Daemon *daemon, const gchar *username, const gcha
 #ifdef HAVE_UTMPX_H
 
 typedef struct {
-        int frequency;
+        guint64 frequency;
         gint64 time;
         GList *previous_logins;
 } UserAccounting;
